requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.3.7
langchain-openai==0.2.9
langchain-ollama==0.2.1
langgraph==0.2.53
pydantic==2.9.2